Include pool_id in pooled sample data
